+++ src/common/components/DatePicker/SlideIcon.tsx
